unfinished refactor
